:wrench: Move to CSS for all components/styles/icons
