+++ packages/fyndiq-component-checkbox/package.json
@@ -11,7 +11,7 @@
   "homepage": "http://developers.fyndiq.com/fyndiq-ui/?selectedKind=Checkbox&selectedStory=default",
   "dependencies": {
     "fyndiq-icon-checkmark": "^1.1.1",
-    "fyndiq-styles-colors": "^1.1.1"
+    "fyndiq-styles-colors": "*"
   },
   "peerDependencies": {
     "react": "^0.14.0 || ^15.0.0"
