+++ packages/fyndiq-component-dropdown/styles.css
@@ -1,4 +1,4 @@
-@import "~fyndiq-styles-colors/colors.less";
+@import "fyndiq-styles-colors/colors.css";
 
 .wrapper {
   display: inline-block;
@@ -13,18 +13,20 @@
   padding: 10px;
   width: auto;
   min-width: 220px;
-  border: 1px solid @bordercolor;
+  border: 1px solid var(--color-border);
   background-color: #FFF;
   border-radius: 3px;
-  box-shadow: 0 3px 8px fadeout(@black, 85%);
+  box-shadow: 0 3px 8px fadeout(var(--color-black), 85%);
 }
 
 .position {
   animation: .15s cubic-bezier(.23, 1, .32, 1.2) normal forwards;
 }
 
-@scaleFactor: .92;
-@translateFactor: 4px;
+:root {
+  --scaleFactor: .92;
+  --translateFactor: 4px;
+}
 
 .position-bl {
   composes: position;
@@ -34,7 +36,7 @@
 @keyframes bl {
   from {
     opacity: 0;
-    transform: translate(0, -@translateFactor) scale(@scaleFactor);
+    transform: translate(0, -var(--translateFactor)) scale(var(--scaleFactor));
   }
 
   to {
@@ -51,7 +53,7 @@
 @keyframes bc {
   from {
     opacity: 0;
-    transform: translate(-50%, -@translateFactor) scale(@scaleFactor);
+    transform: translate(-50%, -var(--translateFactor)) scale(var(--scaleFactor));
   }
 
   to {
@@ -68,7 +70,7 @@
 @keyframes br {
   from {
     opacity: 0;
-    transform: translate(-100%, -@translateFactor) scale(@scaleFactor);
+    transform: translate(-100%, -var(--translateFactor)) scale(var(--scaleFactor));
   }
 
   to {
@@ -77,7 +79,7 @@
   }
 }
 
-// Top
+/* Top */
 
 .position-tl {
   composes: position;
@@ -87,7 +89,7 @@
 @keyframes tl {
   from {
     opacity: 0;
-    transform: translate(0, (-100% + @translateFactor)) scale(@scaleFactor);
+    transform: translate(0, calc(-100% + var(--translateFactor))) scale(var(--scaleFactor));
   }
 
   to {
@@ -104,7 +106,7 @@
 @keyframes tc {
   from {
     opacity: 0;
-    transform: translate(-50%, (-100% + @translateFactor)) scale(@scaleFactor);
+    transform: translate(-50%, calc(-100% + var(--translateFactor))) scale(var(--scaleFactor));
   }
 
   to {
@@ -121,7 +123,7 @@
 @keyframes tr {
   from {
     opacity: 0;
-    transform: translate(-100%, (-100% + @translateFactor)) scale(@scaleFactor);
+    transform: translate(-100%, calc(-100% + var(--translateFactor))) scale(var(--scaleFactor));
   }
 
   to {
