+++ packages/fyndiq-component-price/package.json
@@ -10,7 +10,7 @@
   },
   "homepage": "http://developers.fyndiq.com/fyndiq-ui/?selectedKind=Price&selectedStory=default",
   "dependencies": {
-    "fyndiq-styles-colors": "^1.1.1"
+    "fyndiq-styles-colors": "*"
   },
   "peerDependencies": {
     "react": "^0.14.0 || ^15.0.0"
