+++ packages/fyndiq-component-price/src/currentPrice.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from '../styles.less'
+import styles from '../styles.css'
 
 const CurrentPrice = ({ children }) => (
   <span className={styles.currentPrice}>
