+++ packages/fyndiq-component-price/src/oldPrice.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from '../styles.less'
+import styles from '../styles.css'
 
 const OldPrice = ({ children }) => (
   <span className={styles.oldPrice}>
