+++ packages/fyndiq-component-price/src/price.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from '../styles.less'
+import styles from '../styles.css'
 
 const Price = ({ children }) => (
   <span className={styles.price}>
