+++ packages/fyndiq-component-productcard/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "fyndiq-component-price": "^1.1.1",
     "fyndiq-component-stars": "^1.2.0",
-    "fyndiq-styles-colors": "^1.1.1"
+    "fyndiq-styles-colors": "*"
   },
   "peerDependencies": {
     "react": "^0.14.0 || ^15.0.0"
