--- packages/fyndiq-component-productcard/styles.less
@@ -1,16 +1,0 @@
-@import "~fyndiq-styles-colors/colors.less";
-
-.card {
-  width: 220px;
-  padding: 1em;
-  background: @white;
-  border-radius: .2em;
-}
-
-.image {
-  width: 100%;
-}
-
-.title {
-  color: #333;
-}
