+++ packages/fyndiq-component-productlist/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "fyndiq-component-price": "^1.1.1",
     "fyndiq-component-stars": "^1.2.0",
-    "fyndiq-styles-colors": "^1.1.1",
+    "fyndiq-styles-colors": "*",
     "fyndiq-styles-medias": "*"
   },
   "peerDependencies": {
