+++ packages/fyndiq-component-stars/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Star from 'fyndiq-icon-star'
-import styles from '../styles.less'
+import styles from '../styles.css'
 
 class Stars extends React.Component {
   static propTypes = {
