+++ packages/fyndiq-component-tooltip/styles.css
@@ -1,8 +1,8 @@
-@import "~fyndiq-styles-colors/colors.less";
+@import "fyndiq-styles-colors/colors.css";
 
 .tooltip {
-  background-color: fadeout(@black, 10%);
-  color: @white;
+  background-color: color(var(--color-black) a(90%));
+  color: var(--color-white);
   padding: 3px 5px;
   border-radius: 3px;
   max-width: 200px;
@@ -14,7 +14,7 @@
     height: 0;
     border-style: solid;
     border-width: 0 4px 4px;
-    border-color: transparent transparent @black transparent;
+    border-color: transparent transparent var(--color-black) transparent;
   }
 
   &:empty {
@@ -25,13 +25,13 @@
 .position-b:before {
   top: -4px;
   border-width: 0 4px 4px;
-  border-color: transparent transparent @black transparent;
+  border-color: transparent transparent var(--color-black) transparent;
 }
 
 .position-t:before {
   bottom: -4px;
   border-width: 4px 4px 0;
-  border-color: @black transparent transparent;
+  border-color: var(--color-black) transparent transparent;
 }
 
 .position-bl {
