+++ packages/fyndiq-icon-arrow/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styles from '../styles.less'
+import styles from '../styles.css'
 
 const Arrow = ({ orientation, className }) => {
   let transform
