+++ packages/fyndiq-icon-loader/styles.css
@@ -1,5 +1,3 @@
-@import "~fyndiq-styles-colors/colors.less";
-
 .icon {
   border-radius: 100%;
 }
