+++ packages/fyndiq-icon-star/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styles from '../styles.less'
+import styles from '../styles.css'
 
 const Star = ({ full, onClick, onHover }) => {
   const starId = Math.random() + ''
