--- packages/fyndiq-icon-star/styles.less
@@ -1,19 +1,0 @@
-@import "~fyndiq-styles-colors/colors.less";
-
-.star {
-  display: inline-block;
-  margin-right: .3em;
-  height: 1em;
-  width: 1em;
-  stroke-width: .1em;
-}
-
-.greystar {
-  fill: @lightergrey;
-  stroke: @lightergrey;
-}
-
-.yellowstar {
-  fill: @yellow;
-  stroke: @yellow;
-}
