--- packages/fyndiq-styles-colors/colors.less
@@ -1,18 +1,0 @@
-@red: #EC514A;
-
-@black: #333;
-@darkgrey: #666;
-@lightgrey: #999;
-@lightergrey: #BBB;
-@palegrey: #F2F2F2;
-@white: #FFF;
-
-@blue: #57A5D3;
-@green: #70B65A;
-@yellow: #FFD600;
-@yellow-deal: #FE0;
-@orange: #F08C48;
-
-@bordercolor: #DADADA;
-
-@brand: @red;
