--- packages/fyndiq-styles-fonts/bootstrap.less
@@ -1,6 +1,0 @@
-@import "declarations";
-@import "fonts";
-
-html {
-  font-family: @font-standard;
-}
