+++ packages/fyndiq-styles-fonts/fonts.css
@@ -1,0 +1,4 @@
+:root {
+  --font-standard: 'MuseoSans', Arial, sans-serif;
+  --font-headline: 'Cubano', Arial, sans-serif;
+}
