--- packages/fyndiq-styles-fonts/fonts.less
@@ -1,2 +1,0 @@
-@font-standard: 'MuseoSans', Arial, sans-serif;
-@font-headline: 'Cubano', Arial, sans-serif;
