--- packages/fyndiq-styles-medias/medias.less
@@ -1,10 +1,0 @@
-@screen-xs: 320px;
-@screen-s: 425px;
-@screen-m: 768px;
-@screen-l: 1024px;
-
-@phone-xs: ~"(max-width: @{screen-xs})";
-@phone: ~"(max-width: @{screen-s})";
-@tablet: ~"(max-width: @{screen-m})";
-@desktop: ~"(max-width: @{screen-l})";
-@desktop-l: ~"(min-width: @{screen-l})";
