+++ packages/fyndiq-ui-test/.storybook/webpack.config.js
@@ -4,13 +4,6 @@ const webpack = require('webpack')
 module.exports = {
   module: {
     rules: [{
-      test: /\.less$/,
-      use: [
-        'style-loader',
-        'css-loader?importLoaders=1&sourceMap=true&modules=true',
-        'less-loader',
-      ],
-    }, {
       test: /\.(eot|woff|woff2|ttf)$/,
       use: {
         loader: 'file-loader',
